refactor(frontend): one way for a spec to drive a real grid
The mouse-event drag, the resize by the panel's own corner and the read of
the engine's geometry were spelled out in the browser specs that needed
them, and #87 is a third caller. Two copies of "this is how the library
listens for a drag" is how the specs drift from the library they exist to
guard, so they move to `test/gridEngine.ts` beside the other shared seams.

Waiting for the grid to have been measured is tightened on the way. It
compared the panel against the height of the grid element, which is only
as tall as the rows a page happens to use — on a half-full page that
passes while the panel is still animating to a third of its final size,
and every gesture is then expressed in pixels that mean a different
number of rows than the spec asked for. It now waits for the box the
gesture will actually be given: the row height the engine settled on,
times the rows the panel covers.

That makes the reflow case a genuine two-row drag rather than the one row
it was accidentally performing, and at exactly half a panel's overlap the
engine leaves the two where they are. Dragged by its whole height it is
unambiguous — the top half of a page whose halves are the same size — and
the swap the spec is about is what happens.
